Add a seek bar to set font size scale (#206)

+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/repository/local/android/AndroidSettingsStorage.kt
@@ -36,7 +36,7 @@ class AndroidSettingsStorage(private val androidDatabase: AndroidDatabase) : Loc
         ))
         return@withContext Settings(values.getValue(MetadataDao.KEY_SCREEN_ON).toBoolean(),
                 values.getValue(MetadataDao.KEY_NIGHT_MODE_ON).toBoolean(),
-                values.getValue(MetadataDao.KEY_FONT_SIZE_SCALE).toInt(),
+                values.getValue(MetadataDao.KEY_FONT_SIZE_SCALE).toFloat(),
                 values.getValue(MetadataDao.KEY_SIMPLE_READING_MODE_ON).toBoolean(),
                 values.getValue(MetadataDao.KEY_HIDE_SEARCH_BUTTON).toBoolean(),
                 values.getValue(MetadataDao.KEY_CONSOLIDATE_VERSES_FOR_SHARING).toBoolean(),
